✨ Logout and User model override
It is now possible to logout from a given app by using return app(TrustupIoUserProvider::class)->logout();
This will remove the cookie and redirect the user to auth.trustup.io which will logout the user as well.

Added the ability to override which class is used when the auth()->user() method is called. The class should extend the TrustupIoUserContract.